refactor: 라벨 관련 컴포넌트 이동 (#93)
