+++ src/routes/_authenticated/settings/label/$labelId.tsx
@@ -7,8 +7,8 @@ import { Button, buttonVariants } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { Skeleton } from "@/components/ui/skeleton"
-import { LabelConditionList } from "@/components/label-condition-list"
-import { LabelRuleDialog } from "@/components/label-filter-dialog"
+import { LabelConditionList } from "@/components/label/label-condition-list"
+import { LabelRuleDialog } from "@/components/label/label-filter-dialog"
 import { getErrorMessage } from "@/lib/http-error"
 import { useUpdateLabelRule } from "@/mutations/labels"
 import { useLabelDetail } from "@/queries/labels"
